Move Event and Performance models into results app
