--- phx/performances/tests/models/test_performances_model.py
@@ -1,7 +1,0 @@
-from django.test import TestCase
-
-
-class TestPerformancesModel(TestCase):
-
-    def test_placeholder(self):
-        self.assertEqual(1, 1)
