+++ phx/phx/settings/app.py
@@ -46,8 +46,6 @@ INSTALLED_APPS = [
     'error.apps.ErrorConfig',
     'offline.apps.OfflineConfig',
     'athletes.apps.AthletesConfig',
-    'events.apps.EventsConfig',
-    'performances.apps.PerformancesConfig',
 ]
 
 MIDDLEWARE = [
